fix(inventory): perbaiki mutasi stok & idempotensi penerimaan barang
- FIX CRITICAL: Transfer antar-gudang gagal submit (assert item.quantity, padahal UI kirim item.qty). Normalisasi quantity = item.quantity ?? item.qty di createInventoryTransfer.
- FIX HIGH: goods receipt tidak punya idempotensi -> double-click/retry buat stok & jurnal ganda. Tambah kolom idempotency_key (unique per tenant) + guard di receiveGoods (migration 043).
- FIX HIGH: 6 INSERT stock_movements tidak isi kolom quantity_change (NOT NULL) -> seluruh alur penerimaan/penjualan/refund/servis/micro-component 500. Tambah quantity_change di purchasing, pos.controller, posService, microComponents (dengan cast ::numeric/::integer).
- FIX MEDIUM: addInventoryProduct/updateInventoryProduct/transferProductStock fire-and-forget (syncToApi tanpa rollback). Sekarang await + rollback state + log error bila gagal (polanya sama dgn updateServiceTicket).

Verifikasi: goods receipt idempoten (replay -> 200 duplicate, stok +5 bukan +10, 1 stock_movement); unit 23/23, security 23/23, integration 95/5 (5 = stale regex test), E2E 28/28.

+++ src/server/controllers/purchasing.controller.ts
@@ -326,7 +326,7 @@ export async function receiveGoods(req: Request, res: Response) {
         );
         await client.query(
           `INSERT INTO stock_movements (id, tenant_id, warehouse_id, product_id, type, quantity, quantity_change, reference_no, note)
-           VALUES (gen_random_uuid(), $1, $2, $3, 'PURCHASE_RECEIPT', $4, $4, $5, $6)`,
+           VALUES (gen_random_uuid(), $1, $2, $3, 'PURCHASE_RECEIPT', $4::numeric, $4::integer, $5, $6)`,
           [
             tenantId,
             warehouseId,
